pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ekaterina Konstantinova <feldsher2812@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
postgres = "^4.0"
validators = "^0.20.0"
requests = "^2.31.0"
bs4 = "^0.0.1"
gunicorn = "20.1.0"
markupsafe = "^2.1.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
